>>Feature<< Los productos se pueden añadir por unidad, metros o por kilogramos. Funciona con numeros con decimales

+++ src/main/java/com/example/sistema_ventas/modelo/conexionBD/seleccionarBD/seleccionarBD_cant_stock_id.java
@@ -8,7 +8,7 @@ public class seleccionarBD_cant_stock_id {
 
 
 
-   static public int cantidad(int idProducto) {
+   static public double cantidad(int idProducto) {
 
         conexionBD connectionnow = new conexionBD();
         Connection connectDB = connectionnow.getConnection();
@@ -21,7 +21,7 @@ public class seleccionarBD_cant_stock_id {
             ps.setInt(1, idProducto);
             ResultSet queryOutput = ps.executeQuery();
             queryOutput.next();
-            return queryOutput.getInt("cantidad");
+            return queryOutput.getDouble("cantidad");
 
         } catch (SQLException e) {
             System.out.println(e.getMessage());
